24.12 commit

+++ BlockChainAppMvc/Models/Concrate/Coin.cs
@@ -6,12 +6,13 @@ using System.Threading.Tasks;
 
 namespace BlockChainAppMvc.Models
 {
-    public class Coin :IEntity
+    public class Coin : IEntity
     {
         public int id { get; set; }
         public string coinName { get; set; }
-        public double coinValue { get; set; }
-		public double changePerDay { get; set; }
+        public decimal coinValue { get; set; }
+        public decimal changePerDay { get; set; }
 
-	}
+
+    }
 }
